refactor: Time과 Room 관심사별 객체 분리
