+++ src/main/java/finalmission/controller/TimeController.java
@@ -1,36 +1,19 @@
 package finalmission.controller;
 
-import finalmission.dto.request.RoomCreateRequest;
 import finalmission.dto.request.TimeAddRequest;
-import finalmission.dto.response.RoomCreateResponse;
-import finalmission.dto.response.RoomResponse;
 import finalmission.dto.response.TimeResponses;
 import finalmission.dto.response.TimeStaticsResponses;
 import finalmission.service.NameGenerator;
-import finalmission.service.TimeVoteService;
+import finalmission.service.TimeService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
 @RequiredArgsConstructor
-public class TimeVoteController {
+public class TimeController {
 
     private final NameGenerator nameGenerator;
-    private final TimeVoteService service;
-
-    @PostMapping("/room")
-    public RoomCreateResponse createRoom(
-            @RequestBody RoomCreateRequest request
-    ) {
-        return service.createRoom(request.startDate(), request.endDate(), request.startTime(), request.endTime());
-    }
-
-    @GetMapping("/room/{roomId}")
-    public RoomResponse getRoomInfo(
-            @PathVariable("roomId") String roomId
-    ) {
-        return service.getRoomInfo(roomId);
-    }
+    private final TimeService service;
 
     @PostMapping("/time/{roomId}")
     public void addTime(
