+++ src/main/java/finalmission/service/TimeService.java
@@ -1,39 +1,21 @@
 package finalmission.service;
 
 import finalmission.domain.*;
-import finalmission.dto.response.RoomCreateResponse;
-import finalmission.dto.response.RoomResponse;
 import finalmission.dto.response.TimeResponses;
 import finalmission.dto.response.TimeStaticsResponses;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.time.LocalDate;
 import java.time.LocalDateTime;
-import java.time.LocalTime;
 import java.util.List;
 
 @Service
 @RequiredArgsConstructor
-public class TimeVoteService {
+public class TimeService {
 
     private final RoomRepository roomRepository;
     private final TimeRepository timeRepository;
-
-    @Transactional
-    public RoomCreateResponse createRoom(LocalDate startDate, LocalDate endDate, LocalTime startTime, LocalTime endTime) {
-        Room room = new Room(startDate, endDate, startTime, endTime);
-        roomRepository.save(room);
-        return RoomCreateResponse.from(room);
-    }
-
-    @Transactional(readOnly = true)
-    public RoomResponse getRoomInfo(String roomId) {
-        Room room = roomRepository.findById(new Id(roomId)).orElseThrow();
-
-        return RoomResponse.from(room);
-    }
 
     @Transactional
     public void addTime(String roomId, String username, List<LocalDateTime> values) {
